+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CICD: CI workflow
 - CICD: Changelog workflow
 - Complete API documentation
+- Set precision of time legends to the millisecond
 
 ### Changed